% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColorTable.R
\name{addColorTable}
\alias{addColorTable}
\title{Create a virtual raster file with a color table}
\usage{
addColorTable(x, table)
}
\arguments{
\item{x}{The path to raster file containing categorical data,
must be a single band byte encoded file.}

\item{table}{a  \code{data.frame}  with two or three columns:
\describe{
\item{value}{(numeric) raster cell values.}
\item{color}{(character) the color to assign each value. Typically a
hexadecimal color in #rrggbb format. It will be passed to
\code{\link[grDevices:col2rgb]{grDevices::col2rgb()}} so any format that function accepts will work.}
\item{category}{(optional, character) text describing the class.  If
present this will be used in the map legend.}
}}
}
\value{
\code{addColorTable}  creates a \code{.vrt} file but returns nothing.
}
\description{
\code{addColorTable()}  creates a virtual raster file (\code{.vrt})
alongside a byte encoded raster that references the raster and includes
the color table.  The \code{.vrt} is a GDAL file that describes a
dataset to be composed of other datasets. it can be read directly by
some GIS software or can be passed to \code{\link[=makeNiceTif]{makeNiceTif()}} or GDAL translate
directly to combine the data and color table in a single TIFF file.
}
\details{
Note: The color table will include all values between 0 and the maximum
value in the raster.  If your raster doesn't use some values in the sequence
the missing values will appear in the key with white as the associated color
and an empty string as the category label.
Generally it works best if you recode your raster to have values 0 to N-1,
or 1 to N.

The suggested workflow is:

\enumerate{
\item \code{addColorTable()}   create \code{.vrt} alongside existing  TIFF
\item \code{\link[=makeNiceTif]{makeNiceTif()}}  called directly on the \code{.vrt} to create a TIFF
with embedded color table, tiled data, stats, and overviews.
\item optionally call \code{\link[=addVat]{addVat()}} to build a value attribute table.
}
}
