% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOverviews.R
\name{addOverviews}
\alias{addOverviews}
\title{Function to add overviews to an image file}
\usage{
addOverviews(
  x,
  clean = FALSE,
  compression = "LZW",
  method = "nearest",
  overviews = c(2, 4, 8, 16, 32, 64, 128, 256),
  readonly = TRUE
)
}
\arguments{
\item{x}{The path to a TIFF or other raster file.}

\item{clean}{Defaults to \code{FALSE}; set to \code{TRUE} to remove instead of add
overviews.}

\item{compression}{The style of compression used in the overviews
options are: \code{"LZW"} (the default), \code{"DEFLATE"}, and \code{"JPEG"}.}

\item{method}{(character) the resampling method one of \code{"nearest"},
\code{"average"}, \code{"gauss"}, \code{"cubic"}, \code{"cubicspline"}, \code{"lanczos"},
\code{"average_mp"}, \code{"average_magphase"}, or \code{"mode"}. It is passed to
\href{http://www.gdal.org/gdaladdo.html}{gdaladdo} check there for details.
\code{"near"} is also accepted and silently converted to, \code{"nearest"}.}

\item{overviews}{Overview levels to use.
See \href{https://gdal.org/programs/gdaladdo.html#cmdoption-gdaladdo-arg-levels}{gdaladdo  \if{html}{\out{<levels>}}}}

\item{readonly}{if TRUE \code{x} will not be modified and
overviews will be written to an external file.
See \href{https://gdal.org/programs/gdaladdo.html#cmdoption-gdaladdo-arg-levels}{gdaladdo -ro}}
}
\value{
\code{addOverviews} creates an additional \code{".ovr"} file alongside \code{x}
with overview information.  It does not return anything.
}
\description{
This function creates overviews (aka pyramids) with compressed copies
of the data from a raster at multiple coarser resolutions. They are saved in
a side car file with the extension \code{".ovr"} appended to the original file's
full name (e.g. \code{"dem.tif.ovr"} would contain the overviews for \code{"dem.tif"}).
Overview are used by GIS software to display the raster file more quickly
when zoomed out. They are especially useful for large files.
}
\details{
Generally the default arguments are good but you may want to change the method
to \code{"average"} for continuous raster files and possibly to \code{"mode"} for
categorical files. For continuous files, especially imagery, if you are OK
with lossy compression in the overviews you may change the compression to
\code{"JPEG"} which should result in smaller files.

\code{addOveriews} is a convenience wrapper to the
\href{https://www.gdal.org/gdaladdo.html}{gdaladdo} utility.
\code{addOverviews()} sets reasonable defaults, only some of which can be
overridden.
}
