% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyTif.R
\name{copyTif}
\alias{copyTif}
\title{Copy a TIFF and all its associated files}
\usage{
copyTif(from, to, includeVrt = TRUE, overwrite = FALSE)
}
\arguments{
\item{from}{(character) path to a TIFF to be copied}

\item{to}{(character) path where copy will be created}

\item{includeVrt}{(logical) if \code{TRUE} also copy an associated \code{.vrt} file
with same base name.}

\item{overwrite}{(logical) if \code{TRUE} overwrite existing file if \code{FALSE}
don't.}
}
\value{
Nothing is returned.
}
\description{
Copy a TIFF and all its associated files
}
