% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeToReference.R
\name{rasterizeToReference}
\alias{rasterizeToReference}
\title{Function to rasterize vector data}
\usage{
rasterizeToReference(
  source,
  destination,
  reference,
  burn,
  attribute,
  init,
  type = "Byte",
  allTouched = FALSE,
  sql
)
}
\arguments{
\item{source}{(character) path to a vector GIS file (such as a shapefile)
that contains data to be rasterized.}

\item{destination}{(character) path to a \code{.tif} file that will either be created or added to.}

\item{reference}{(character) path to raster dataset will be used to define extent, cellsize, and projection
used to define the extent of destination if destination does not already exist.}

\item{burn}{(optional, numeric) value to be burnt into grid (everywhere the \code{source} has data)}

\item{attribute}{(optional, character) the column name in the source to extract values from}

\item{init}{(optional, numeric) if supplied this value will be written into the grid as a background value
prior to rasterizing.  init can only be used if the destination does not already exist.}

\item{type}{the type of grid to create. Should be one of \code{"Byte"}, \code{"UInt16"},
\code{"Int16"}, \code{"UInt32"}, \code{"Int32"}, \code{"Float32"},
\code{"Float64"} or for convenience you may
also use the raster package's designations: \code{\link[raster]{dataType}}.}

\item{allTouched}{(optional, logical) defaults to FALSE.  If TRUE
"all pixels touched by lines or polygons will be updated, not just those on
the line render path, or whose center point is within the polygon"
see "-at" flag description for gdal_rasterize: \url{http://www.gdal.org/gdal_rasterize.html}}

\item{sql}{(optional, character) An SQL statement to be applied to the data
source.  This can be used to filter or sort the data prior to burning.}
}
\value{
This function creates a new raster or writes values to an existing raster
 at the destination.  It does not return anything.
}
\description{
This function rasterizes vector data into a file that coincides (same extent,
cellsize, projection, and cell alignment) with a reference raster file.
}
\details{
If the file at \code{destination} exists the new pixels will be overlaid on
existing data such that existing data will persist where it does not underlie
the vector data in \code{source}.

Either the \code{burn} or \code{attribute} argument should be used to set the
value assigned to cells.  The default type (\code{"Byte"}) supports integers 0 to
255.  If you are writing real numbers or outside
of that range be sure to set \code{type} to something appropriate.  Similarly,
if you are only writing a few values you may want to set \code{type} to something
that supports fewer values such as \code{"Byte"}.

When creating a new file
the CRS of the source must match that of the reference and this used to be checked
with \code{\link[raster]{compareCRS}} but with updates to the proj library and the
switch to WKT representations I don't currently have a good way to check if two
projections are equivallent. Its now up to the user to make sure that they are.

 In contrast if you are overlaying data on an existing file then the underlying gdal_rasterize
function will reproject while rasterizing.
}
