% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPrepOptions.R
\name{rasterPrepOptions}
\alias{rasterPrepOptions}
\title{Function to change rasterPrep package settings}
\usage{
rasterPrepOptions(...)
}
\arguments{
\item{...}{Arguments should be settings to reset with their value the new setting.
If called with no arguments nothing is changed but the current settings are printed.}
}
\value{

}
\description{
This functions allows the user to print or change the settings for the rasterPrep package
}
\details{
Currently there are two settings:

setProjLib -  controls whether raster prep should attempt to set
the system environmental setting "PROJ_LIB" before executing system and gdalUtils calls
it is initially TRUE.

projLib - is what the PROJ_LIB environental setting should be set to, intitially "", an empty string.

On my system in a clean R session Sys.getenv("PROJ_LIB") will return "" but
after rgdal is loaded it will be "C:/Users/user/Documents/R/win-library/4.0/rgdal/proj"
Unfortunately that means that when you try to run gdal_utulities via a system call
they don't use the co-installed proj library but instead use rgdal's version.
If the versions match this may not be a problem but if they don't things may go wrong
in sometimes obvious and sometimes very subtle ways.

With the intial values
rasterPrep will change the PROJ_LIB enviromental setting to "" before making a system call.

This can be turned off with rasterPrepOptions(setProjLib = FALSE) or changed to
a different proj lib location with rasterPrepOptions(projLib = "C:/Proj/") or some other path
appropriate for your system.

The package will always reset the PROJ_LIB value to the original value after the system call
so subsequent  rgdal functions will still find it set as expected by that package.

See:   https://github.com/r-spatial/discuss/issues/31 for a discussion of this issue.
}
