% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameTif.R
\name{renameTif}
\alias{renameTif}
\title{Function to rename or move a .tif and any associated sidecar files}
\usage{
renameTif(from, to, includeVrt = TRUE)
}
\arguments{
\item{from}{(character) path to a tif to be renamed}

\item{to}{(character) path to renamed tif}

\item{includeVrt}{(logical) if TRUE also rename an associated .vrt file with
same base name.}
}
\value{
this function returns nothing
}
\description{
Function to rename or move a .tif and any associated sidecar files
}
