% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTifFiles.R
\name{findTifFiles}
\alias{findTifFiles}
\title{function to find all the files associated with a .tif file.}
\usage{
findTifFiles(x, includeVrt = TRUE)
}
\arguments{
\item{x}{(character) the path to a single .tif file}

\item{includeVrt}{(logical) if TRUE consider a .vrt with the same base name
to be associated.}
}
\value{
a vector of files associated with x
}
\description{
This is primarily intended for internal use and is called by
\code{\link{copyTif}},\code{\link{deleteTif}}, and \code{\link{renameTif}}.
}
\details{
Given the path to a .tif file return the path to all the files associated with
it including the original .tif.  Any file that includes the full name (with
.tif extension and some additional extension  e.g. .tif.orv) a world
file which has the same base name but .twf instead of .tif and if
\code{includeVrt} is TRUE a .vrt file with the same name.
}
\keyword{internal}
